Modificando main

+++ AirBender.cpp
@@ -4,7 +4,7 @@ AirBender::AirBender(){
 
 }
 
-AirBender::AirBender(int defensa, int ofensa, int fuerza, int suerte) : Bender(defensa, ofensa, fuerza, suerte){
+AirBender::AirBender(double defensa, double ofensa, double fuerza, double suerte) : Bender(defensa, ofensa, fuerza, suerte){
 
 }
 
@@ -14,7 +14,7 @@ void AirBender::Ataque(Bender* bender){
 
 void AirBender::AtaqueEspecial(Bender* bender){
   if(typeid(*bender).name() == typeid(FireBender).name()){
-
+    
   }
   if(typeid(*bender).name() == typeid(EarthBender).name()){
