+++ FireBender.cpp
@@ -4,7 +4,7 @@ FireBender::FireBender(){
 
 }
 
-FireBender::FireBender(int defensa, int ofensa, int fuerza, int suerte) : Bender(defensa, ofensa, fuerza, suerte){
+FireBender::FireBender(double defensa, double ofensa, double fuerza, double suerte) : Bender(defensa, ofensa, fuerza, suerte){
 
 }
